+++ apps/hello-work-job-searcher/src/app/client/hooks/useInfiniteJobList.ts
@@ -1,4 +1,4 @@
-import { type TJobOverview, jobListSuccessResponseSchema } from "@sho/models";
+import { jobListSuccessResponseSchema, type TJobOverview } from "@sho/models";
 import { useInfiniteQuery } from "@tanstack/react-query";
 
 async function fetchServerPage(nextToken?: string) {
