+++ apps/hello-work-job-searcher/src/app/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import "./styles/globals.css";
-import { QueryProvider } from "./components/client/QueryProvider";
+import { QueryProvider } from "./client/components/QueryProvider";
 export const metadata: Metadata = {
   title: "ハローワーク求人情報検索",
   description: "Generated by create next app",
